% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edwards.R
\name{edwards}
\alias{edwards}
\title{edwards}
\source{
Edwards, M Richardson A.J. Impact of climate change on marine pelagic
phenology and trophic mismatch. Nature, 2004, 430, 881-884.
\url{http://dx.doi.org/10.1038/nature02808}.
}
\usage{
edwards(value, time)
}
\arguments{
\item{value}{a numerical vector of values (abundances...)}

\item{time}{a numerical vector of temporal units (month, week...)}
}
\value{
a numerical value corresponding to the timing of the seasonal peak in time unit corresponding the time vector
}
\description{
This function computes the timing of a seasonal peak using Edwards and Richardson formula
}
\section{Details}{

The seasonal peak timing is defined throughout the entire growing season using
the time scale coordinate (number of the month, week, day...) of the centre of
gravity of the area below the plot of the corresponding values.


For monthly data the seasonal peak timing \eqn{T} is:
\deqn{T= \frac{\sum_{i=1}^{12}M.x_m}{\sum_{i=1}^{12}x_m}}
with \eqn{x_m} is the abundance in month \eqn{M} (January=1,...Decembre=12).
}

\examples{
\dontrun{
time<-1:12
value<-c(0,0,1,2,5,7,3,0,0,0,0,0)
edwards(time,value)

}
}
\author{
Laurent Dubroca
}
