% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytopknar.R
\docType{data}
\name{phytopknar}
\alias{phytopknar}
\title{phytopknar}
\format{
a dataframe with 1841 observations and 166 variables

Columns:

location = sampling location

station = station name

date = date of the sampling in Date format

column 4 to 1666 = species abundances in cell.mL-1. The species latin name is given
by the column name.
}
\source{
Smayda, T.J. & the Bunker C community (1959-1997).  Narragansett Bay
Plankton Time Series.  Graduate School of Oceanography, URI.  Data available
at:  \url{www.NABATS.org}. Preprocessing by Lorenzo Longobardi.
}
\usage{
phytopknar
}
\description{
Narragansett bay time series of phytoplakton abundances
}
\details{
Methods: Samples collected at Narragansett Bay, RI, US  "Station II", Latitude 41° 34' 2" Longitude 71° 23' 4" and counted live 1-3 hours after collection.
Whole-water (unconcentrated) counts done with a Sedgwick-Rafter cell at 250X and 500X magnification, phase contrast optics.  Entire chamber counted for 1 cell /mL detection level.
UNITS ARE CELLS /mL. Phytoplankton species names reflect those used in the early 2000s, and have not been updated to current taxonomy.
}
\keyword{datasets}
