% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ret_time.R
\name{ret_time}
\alias{ret_time}
\title{ret_time}
\usage{
ret_time(x)
}
\arguments{
\item{x}{a dataframe containing a as.Date object named as "date"}
}
\value{
a dataframe containing the time vectors (year, month...)
}
\description{
Return time vectors from a dataframe containing an as.Date object
}
\examples{
\dontrun{
  date <- seq(as.Date("2000/1/1"), as.Date("2001/1/1"),length.out=10)
   value <- c(seq(1,10,1))
   df <- data.frame(date,value)
   ret_time(df)
}
}
