% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_multi.R
\name{spline_multi}
\alias{spline_multi}
\title{spline_multi}
\usage{
<<<<<<< HEAD
spline_multi(x, ab_treshold = 0.75, obs_year = 2, s_param = 0.35,
  t_scale = 52)
=======
spline_multi(x, ab_treshold = 0.75, obs_year = 2, s_param = 0.35, t_scale = 52)
>>>>>>> 46527344f37f3888c33ddded3dd15688f9b7da3e
}
\arguments{
\item{x}{a dataframe having the first 7 columns named as "location", "station", "date", "year", "month", "week", "day",
other columns corresponds to the species columns}

\item{ab_treshold}{quantile of the positive abundance distribution (default 0.75)}

\item{obs_year}{minimum number of samples for each year in which the species shows an abundance>0 (default 2)}

\item{s_param}{the smoothing parameter corresponding to the spar parameters
of the smooth.spline function (default 0.35).}

\item{t_scale}{temporal scale resolution of the data, 52 for weekly data, 12 for monthly data (default 52)}
}
\value{
a dataframe containing the species names, the annual time vector, the value vector and info
a character vector with 'Start', 'Max' and 'End' for the seasonal peak and NA
for the other date
}
\description{
Identification of annual bloom descriptors dates based on smoothing splines for all the species of the dataframe.
<<<<<<< HEAD
The function automatically exludes those years containing less than  75% of
=======
The function automatically exludes those years containing less than  75\% of
>>>>>>> 46527344f37f3888c33ddded3dd15688f9b7da3e
the possible observations for the temporal scale considered.
}
\examples{
\dontrun{
  data("phytopknar")
   phytopknar_ret <- ret_time(phytopknar)
   phytopknar_ret_ord <- phytopknar_ret \%>\% dplyr::select(location,station,date,year,month,week,day,everything())
   spline_multi(phytopknar_ret_ord,ab_treshold=0.75,obs_year=2,s_param=0.35,t_scale=52)
}
}
