% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_points.R
\name{spline_points}
\alias{spline_points}
\title{Identification of bloom descriptors dates based on smoothing splines}
\usage{
spline_points(x, s_param = 0.35)
}
\arguments{
\item{x}{A dataframe containing the temporal information and the abundance of a species.}

\item{s_param}{The smoothing parameter (default 0.35).}
}
\value{
A new dataframe containing temporal and species abundance info and date of start, maximum abundance and end of the bloom of the species.
}
\description{
Identification of bloom descriptors dates based on smoothing splines
}
\examples{
\dontrun{
spline.points(x,s_param=0.35)
}
}
