% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_points.R
\name{spline_points}
\alias{spline_points}
\title{spline_points}
\usage{
spline_points(value, time, s_param = NULL, control = NULL, past = NULL)
}
\arguments{
\item{value}{a numerical vector of values (abundances...)}

\item{time}{a numerical vector of temporal units (month, week...)}

\item{s_param}{the smoothing parameter corresponding to the spar parameters
of the smooth.spline function (default 0.35).}

\item{control}{control the flexibility of the function (default 0).}

\item{past}{number of observations before the the maximum derivative. Default
0, other values are useful whether the user wants to pick the moment just before
the max derivative (max growth).}
}
\value{
a dataframe containing the time vector, the value vector and info
a character vector with 'Start', 'Max' and 'End' for the seasonal peak and NA
for the other date
}
\description{
Identification of bloom descriptors dates based on smoothing splines
}
\examples{
\dontrun{
time <- 1:12
value <- c(0, 0, 1, 2, 5, 7, 3, 0, 0, 0, 0, 0)
spline_points(value, time, s_param = 0.35, control = 0, past = 0)
}
}
